Improve update avatar function

+++ src/public/data/changeAvatar.php
@@ -12,6 +12,7 @@
 
   $json = new Services_JSON();
   $userId = $_POST['userId'];
+  $uploaded = false;
 
   $baseUrl = 'users/';
   $imgDir = 'data/images/';
@@ -22,20 +23,18 @@
     $temp = explode(".", $_FILES["file"]["name"]);
     $newfilename =  str_replace(' ', '', $user->firstname . $user->lastname) . '-' . date('Y-m-d') . '.' . end($temp);
 
-    if (file_exists($user->avatar)) {
-      unlink($user->avatar);
-    }
-
-    move_uploaded_file($_FILES["file"]["tmp_name"], 'images/' . $newfilename);
-
-    if (file_exists($baseUrl . $userId . '.json')) {
-      $user->avatar = $imgDir . $newfilename;
-      unlink($baseUrl . $userId . '.json');
-      file_put_contents($baseUrl . $userId . ".json", $json->encode($user));
-      echo $json->encode($user);
+    if (move_uploaded_file($_FILES["file"]["tmp_name"], 'images/' . $newfilename)) {
+      $uploaded = true;
+      if (file_exists($baseUrl . $userId . '.json')) {
+        $user->avatar = $imgDir . $newfilename;
+        unlink($baseUrl . $userId . '.json');
+        file_put_contents($baseUrl . $userId . ".json", $json->encode($user));
+        echo $json->encode(array('uploaded' => $uploaded, 'user' => $user));
+      } else {
+        echo $json->encode(array('uploaded' => $uploaded, 'user' => array()));
+      }
     } else {
-      echo $json->encode(array());
+      echo $json->encode(array('uploaded' => $uploaded, 'user' => array()));
     }
   }
-
 ?>
